+++ HttpStatusCodes_C.h
@@ -104,7 +104,8 @@ enum HttpStatus_Code
 	HttpStatus_VariantAlsoNegotiates         = 506, /*!< Indicates that the server has an internal configuration error: the chosen variant resource is configured to engage in transparent content negotiation itself, and is therefore not a proper end point in the negotiation process. */
 	HttpStatus_InsufficientStorage           = 507, /*!< Means the method could not be performed on the resource because the server is unable to store the representation needed to successfully complete the request. */
 	HttpStatus_LoopDetected                  = 508, /*!< Indicates that the server terminated an operation because it encountered an infinite loop while processing a request with "Depth: infinity". [RFC 5842] */
-	HttpStatus_NotExtended                   = 510, /*!< The policy for accessing the resource has not been met in the request. [RFC 2774] */
+	HttpStatus_NotExtended                   = 510, /*!< \deprecated \parblock Obsoleted as the experiment has ended and there is no evidence of widespread use. \endparblock
+	                                                     The policy for accessing the resource has not been met in the request. [RFC 2774] */
 	HttpStatus_NetworkAuthenticationRequired = 511, /*!< Indicates that the client needs to authenticate to gain network access. */
 
 	HttpStatus_xxx_max = 1023